+++ tonic-build/src/client.rs
@@ -44,6 +44,7 @@ pub fn generate<T: Service>(
                 clippy::let_unit_value,
             )]
             use tonic::codegen::*;
+            use tonic::codegen::http::Uri;
 
             #service_doc
             #(#struct_attributes)*
@@ -58,7 +59,7 @@ pub fn generate<T: Service>(
             where
                 T: tonic::client::GrpcService<tonic::body::BoxBody>,
                 T::Error: Into<StdError>,
-                T::ResponseBody: Default + Body<Data = Bytes> + Send  + 'static,
+                T::ResponseBody: Body<Data = Bytes> + Send  + 'static,
                 <T::ResponseBody as Body>::Error: Into<StdError> + Send,
             {
                 pub fn new(inner: T) -> Self {
@@ -66,9 +67,15 @@ pub fn generate<T: Service>(
                     Self { inner }
                 }
 
+                pub fn with_origin(inner: T, origin: Uri) -> Self {
+                    let inner = tonic::client::Grpc::with_origin(inner, origin);
+                    Self { inner }
+                }
+
                 pub fn with_interceptor<F>(inner: T, interceptor: F) -> #service_ident<InterceptedService<T, F>>
                 where
                     F: tonic::service::Interceptor,
+                    T::ResponseBody: Default,
                     T: tonic::codegen::Service<
                         http::Request<tonic::body::BoxBody>,
                         Response = http::Response<<T as tonic::client::GrpcService<tonic::body::BoxBody>>::ResponseBody>
@@ -78,20 +85,20 @@ pub fn generate<T: Service>(
                     #service_ident::new(InterceptedService::new(inner, interceptor))
                 }
 
-                /// Compress requests with `gzip`.
+                /// Compress requests with the given encoding.
                 ///
                 /// This requires the server to support it otherwise it might respond with an
                 /// error.
                 #[must_use]
-                pub fn send_gzip(mut self) -> Self {
-                    self.inner = self.inner.send_gzip();
+                pub fn send_compressed(mut self, encoding: CompressionEncoding) -> Self {
+                    self.inner = self.inner.send_compressed(encoding);
                     self
                 }
 
-                /// Enable decompressing responses with `gzip`.
+                /// Enable decompressing responses.
                 #[must_use]
-                pub fn accept_gzip(mut self) -> Self {
-                    self.inner = self.inner.accept_gzip();
+                pub fn accept_compressed(mut self, encoding: CompressionEncoding) -> Self {
+                    self.inner = self.inner.accept_compressed(encoding);
                     self
                 }
 
@@ -166,7 +173,7 @@ fn generate_unary<T: Method>(
     compile_well_known_types: bool,
     path: String,
 ) -> TokenStream {
-    let codec_name = syn::parse_str::<syn::Path>(T::CODEC_PATH).unwrap();
+    let codec_name = syn::parse_str::<syn::Path>(method.codec_path()).unwrap();
     let ident = format_ident!("{}", method.name());
     let (request, response) = method.request_response_name(proto_path, compile_well_known_types);
 
@@ -191,7 +198,7 @@ fn generate_server_streaming<T: Method>(
     compile_well_known_types: bool,
     path: String,
 ) -> TokenStream {
-    let codec_name = syn::parse_str::<syn::Path>(T::CODEC_PATH).unwrap();
+    let codec_name = syn::parse_str::<syn::Path>(method.codec_path()).unwrap();
     let ident = format_ident!("{}", method.name());
 
     let (request, response) = method.request_response_name(proto_path, compile_well_known_types);
@@ -217,7 +224,7 @@ fn generate_client_streaming<T: Method>(
     compile_well_known_types: bool,
     path: String,
 ) -> TokenStream {
-    let codec_name = syn::parse_str::<syn::Path>(T::CODEC_PATH).unwrap();
+    let codec_name = syn::parse_str::<syn::Path>(method.codec_path()).unwrap();
     let ident = format_ident!("{}", method.name());
 
     let (request, response) = method.request_response_name(proto_path, compile_well_known_types);
@@ -243,7 +250,7 @@ fn generate_streaming<T: Method>(
     compile_well_known_types: bool,
     path: String,
 ) -> TokenStream {
-    let codec_name = syn::parse_str::<syn::Path>(T::CODEC_PATH).unwrap();
+    let codec_name = syn::parse_str::<syn::Path>(method.codec_path()).unwrap();
     let ident = format_ident!("{}", method.name());
 
     let (request, response) = method.request_response_name(proto_path, compile_well_known_types);