+++ tonic/src/transport/channel/mod.rs
@@ -107,7 +107,7 @@ impl Channel {
 
     /// Balance a list of [`Endpoint`]'s.
     ///
-    /// This creates a [`Channel`] that will load balance accross all the
+    /// This creates a [`Channel`] that will load balance across all the
     /// provided endpoints.
     pub fn balance_list(list: impl Iterator<Item = Endpoint>) -> Self {
         let (channel, tx) = Self::balance_channel(DEFAULT_BUFFER_SIZE);